chore: seperate internal node functions into "basis.ts"
