+++ src/nodes/SamplerNode.tsx
@@ -4,6 +4,8 @@ import { memo, useEffect, useState } from "react";
 import { RiVoiceprintFill } from "@remixicon/react";
 
 import { makeAsyncNodeFactory, NodeTypeDescriptor } from ".";
+import type { NodeTypeDescriptor } from ".";
+import { makeNodeFactory } from "./basis";
 import { AudioGenerator, NoteEvent, InstrumentNodeData, NOTE_INPUT_HID_MAIN, SIGNAL_OUTPUT_HID, AudioDestination } from "../graph";
 import { Automatable } from "../parameters";
 import { NodeDataSerializer } from "../serializer";
